update npmignore for ts build
With the introduction of typescript we need to change the exclude
list of the package generation.

This commit does just that, by flipping the logic to an include list.
Maintaining an include list is actually easier in our case than the
exclude list, but we can't use the "files" package.json field as
we need to negate specific patterns in public/.
Patterns like `!public/js/bundle.vendor.js.map` in "files" make `yarn
pack` include all the files, as the check they do is buggy.
`npm pack` would work fine, but then we mix package managers again.

Using the allow list in .npmignore makes it work with both, npm and
yarn, but gets a bit more tedious to read for a human.